new api directories, fixed imports
