+++ pos_back_end/api/admin/admin_request_models.py
@@ -1,20 +1,12 @@
 from pydantic import BaseModel, Field, ConfigDict
 
+from pos_back_end.api.address.address_models import AddressRequestAndResponse
 from pos_back_end.db.models.address import Address
 
 
 class LoginRequestBody(BaseModel):
     email: str
     password: str
-
-
-class AddressRequestAndResponse(BaseModel):
-    street_address: str
-    city: str
-    state: str
-    zipcode: str
-
-    model_config = ConfigDict(from_attributes=True)
 
 
 class LoginResponseBody(BaseModel):
