+++ pos_back_end/api/admin/admin_requests.py
@@ -2,50 +2,14 @@ from fastapi import APIRouter, HTTPException, status
 from fastapi import Depends
 from sqlalchemy.orm import Session
 from pos_back_end.db.dependencies import get_db
-from pos_back_end.api.admin.models.admin_request_models import LoginRequestBody, LoginResponseBody, PostAdminRequestBody, PostAdminResponseBody, \
+from pos_back_end.api.admin.admin_request_models import PostAdminRequestBody, PostAdminResponseBody, \
     AddressRequestAndResponse
-from .models.state_models import StateResponseBody, StateItem
-from .services.address_services import AddressServices
-from .services.admin_services import AdminServices
-from .services.login_services import LoginServices
+from pos_back_end.api.address.address_services import AddressServices
+from pos_back_end.api.admin.admin_services import AdminServices
+from pos_back_end.api.login.login_services import LoginServices
 from ...db.models.admin import Admin
 
 router = APIRouter()
-
-
-@router.post("/login")
-def login(request: LoginRequestBody, db: Session = Depends(get_db)):
-
-    service = LoginServices()
-
-    admin: Admin = service.get_admin_data(
-        request.email,
-        request.password,
-        db
-    )
-
-    if not admin:
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Invalid email or password",
-        )
-
-    address = admin.address
-
-    return LoginResponseBody(
-        user_id=admin.id,
-        first_name=admin.first_name,
-        last_name=admin.last_name,
-        company_name=admin.company_name,
-        email=admin.email,
-        phone_number=admin.phone_number,
-        address=AddressRequestAndResponse(
-            street_address=address.street_address,
-            city=address.city,
-            state=address.state,
-            zipcode=address.zipcode
-        ) if address is not None else None
-    )
 
 
 @router.post("/admin")
@@ -92,11 +56,3 @@ def create_admin(request: PostAdminRequestBody, db: Session = Depends(get_db)):
     )
 
 
-@router.get("/states")
-def get_states(db: Session = Depends(get_db)):
-
-    states = AddressServices().get_states(db)
-
-    state_items = [StateItem.from_orm(state) for state in states]
-
-    return StateResponseBody(states=state_items)
