+++ pos_back_end/api/login/__init__.py
@@ -1,0 +1,6 @@
+
+from fastapi import APIRouter
+from .login_requests import router as get_login_router
+
+router = APIRouter()
+router.include_router(get_login_router)
