+++ pos_back_end/api/login/login_services.py
@@ -1,7 +1,9 @@
+from fastapi import APIRouter
 from sqlalchemy.orm import Session
-
-from pos_back_end.api.admin.models.admin_request_models import PostAdminRequestBody
 from pos_back_end.db.models.admin import Admin
+
+
+router = APIRouter()
 
 
 class LoginServices:
