+++ pos_back_end/api/restaurant/__init__.py
@@ -1,0 +1,6 @@
+
+from fastapi import APIRouter
+from .restaraunt_requests import router as get_restaurant_router
+
+router = APIRouter()
+router.include_router(get_restaurant_router)
